Added authentice_user for tailgate controller and updated logic only allow users who created tailgate to update or destroy

+++ app/controllers/tailgates_controller.rb
@@ -1,4 +1,6 @@
 class TailgatesController < ApplicationController
+  before_action :authenticate_user, only: [:create, :update, :destroy]
+
   def index
     tailgates = Tailgate.all
 
@@ -36,34 +38,44 @@ class TailgatesController < ApplicationController
 
   def show
     tailgate = Tailgate.find_by(id: params[:id])
+
     render json: tailgate
   end
 
   def update
     tailgate = Tailgate.find_by(id: params[:id])
 
-    tailgate.name = params[:name] || tailgate.name
-    tailgate.description = params[:description] || tailgate.description
-    tailgate.address = params[:address] || tailgate.address
-    tailgate.start_time = params[:start_time] || tailgate.start_time
-    tailgate.end_time = params[:end_time] || tailgate.end_time
-    tailgate.parking_available = params[:parking_available] || tailgate.parking_available
-    tailgate.private_bathroom = params[:private_bathroom] || tailgate.private_bathroom
-    tailgate.alcohol_allowed = params[:alcohol_allowed] || tailgate.alcohol_allowed
-    tailgate.food_provided = params[:food_provided] || tailgate.food_provided
-    tailgate.family_friendly = params[:family_friendly] || tailgate.family_friendly
-    tailgate.tv_available = params[:tv_available] || tailgate.tv_available
+    if tailgate.user_id == current_user.id
+      tailgate.name = params[:name] || tailgate.name
+      tailgate.description = params[:description] || tailgate.description
+      tailgate.address = params[:address] || tailgate.address
+      tailgate.start_time = params[:start_time] || tailgate.start_time
+      tailgate.end_time = params[:end_time] || tailgate.end_time
+      tailgate.parking_available = params[:parking_available] || tailgate.parking_available
+      tailgate.private_bathroom = params[:private_bathroom] || tailgate.private_bathroom
+      tailgate.alcohol_allowed = params[:alcohol_allowed] || tailgate.alcohol_allowed
+      tailgate.food_provided = params[:food_provided] || tailgate.food_provided
+      tailgate.family_friendly = params[:family_friendly] || tailgate.family_friendly
+      tailgate.tv_available = params[:tv_available] || tailgate.tv_available
 
-    if tailgate.save
-      render json: { message: "Tailgate updated successfully", tailgate: tailgate }, status: :accepted
+      if tailgate.save
+        render json: { message: "Tailgate updated successfully", tailgate: tailgate }, status: :accepted
+      end
     else
-      render json: { errors: tailgate.errors.full_messages }, status: :unprocessable_entity
+      render json: { message: "User did not create this post and therefore cannot update it" }, status: :bad_request
     end
   end
 
   def destroy
     tailgate = Tailgate.find_by(id: params[:id])
-    tailgate.destroy
-    render json: { message: "Product successfully destroyed." }
+
+    if tailgate.user_id == current_user.id && DateTime.now < tailgate.start_time
+      tailgate.destroy
+      render json: { message: "Tailgate successfully destroyed" }
+    elsif tailgate.user_id != current_user.id
+      render json: { message: "User did not create this post and therefore cannot update it" }, status: :bad_request
+    else
+      render json: { message: "Cannot delete this tailgate as it has already happened" }
+    end
   end
 end
